requirements.txt
python-docx-1.1.0